Boilerplate

Documentation

data Bit

Constructors

C

S

notBit :: Bit -> Bit

eqBit :: Bit -> Bit -> Bit

andBit :: Bit -> Bit -> Bit

orBit :: Bit -> Bit -> Bit

xorBit :: Bit -> Bit -> Bit

plusCBit :: Bit -> Bit -> Bit -> (Bit, Bit)

minusCBit :: Bit -> Bit -> Bit -> (Bit, Bit)

data W8

Constructors

W8 Bit Bit Bit Bit Bit Bit Bit Bit

zeroW8 :: W8

oneW8 :: W8

notW8 :: W8 -> W8

andW8 :: W8 -> W8 -> W8

xorW8 :: W8 -> W8 -> W8

eqW8 :: W8 -> W8 -> Bit

rolW8 :: W8 -> W8

rorW8 :: W8 -> W8

plusCW8 :: W8 -> W8 -> Bit -> (Bit, W8)

data W32

Constructors

W32 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

w32_0 :: W32

w32_4 :: W32

w32_8 :: W32

plusW32 :: W32 -> W32 -> Bit -> W32

plusCW32 :: W32 -> W32 -> Bit -> (Bit, W32)

andW32 :: W32 -> W32 -> W32

orW32 :: W32 -> W32 -> W32

plusW8 :: W8 -> W8 -> (Bit, W8)

negW8 :: W8 -> W8

minusCW8 :: W8 -> W8 -> Bit -> (Bit, W8)

shlCW8 :: W8 -> Bit -> (Bit, W8)

shrCW8 :: W8 -> Bit -> (Bit, W8)

msbW8 :: W8 -> Bit

lsbW8 :: W8 -> Bit

(<>) :: Monad m => (a -> m b) -> (b -> m c) -> a -> m c

top6 :: W32 -> W6

byte0 :: W32 -> W8

byte1 :: W32 -> W8

byte2 :: W32 -> W8

byte3 :: W32 -> W8

data W5

Constructors

W5 Bit Bit Bit Bit Bit

data W6

Constructors

W6 Bit Bit Bit Bit Bit Bit

data W11

Constructors

W11 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

data W16

Constructors

W16 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

data W26

Constructors

W26 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

zero16 :: W16

zero32 :: W32

one32 :: W32

(||) :: W16 -> W16 -> W32

w32_lte :: W32 -> W32 -> Bool

w32_lt :: W32 -> W32 -> Bool

w32_ne :: W32 -> W32 -> Bool