Boilerplate.Instruction

Documentation

data Instruction l sl r m a

The definition of some abstract instruction

Constructors

Ins

Fields

name :: String

form :: ArgumentsForm l

exec :: ExecutionFunction l sl r m a

data Argument l r

Constructors

Register l

Immediate r

data Arguments l r

arguments to a command represent a mapping of register argument labels to their register label or immediate data example: ra -> r1 or rd -> r32, imm -> 123

Constructors

Args (Map l (Argument l r))

newtype ArgumentsForm l

Constructors

Form [(ArgumentForm, l)]

data ArgumentForm

Constructors

Reg

Imm

reg :: l -> ArgumentsForm l

imm :: l -> ArgumentsForm l

getArg :: Ord l => Arguments l r -> l -> Maybe (Argument l r)

unpackArgs :: Ord l => Arguments l r -> [l] -> Maybe [Argument l r]

formToList :: ArgumentsForm l -> [l]

unpackFromForm :: Ord l => ArgumentsForm l -> Arguments l r -> Maybe [Argument l r]

data Command l sl r m a

the command that actually executes on a machine

Constructors

Command

Fields

instruction :: Instruction l sl r m a

arguments :: Arguments l r

getExec :: Command l sl r m a -> ExecutionFunction l sl r m a

getArgs :: Command l sl r m a -> Arguments l r

unpackCommand :: Command l sl r m a -> (Arguments l r, ExecutionFunction l sl r m a)

type ExecutionFunction l sl r m a = ReaderT (Arguments l r) (MachineST l sl r m) a

function that executes on the state given a set of arguments

type InstructionSet l sl r m a = Map String (Instruction l sl r m a)

execute :: Command l sl r m a -> MachineST l sl r m a

argsFromList :: Ord l => [(l, Argument l r)] -> Arguments l r

setFromList :: [Instruction l sl r m a] -> InstructionSet l sl r m a