Boilerplate.Machines.SpecialPurposeRegister

Bit Configuration

The BitConfiguration type represents a naming of bits in a MachineStatus Register

This is used to generate utility functions to access and manipulate individual bits with the notion that they are flags representing different aspects of the instruction set such as carry, overflow, division by zero, etc.

newtype BitConfiguration l

A naming configuration for bits in a MachineStatus Register. This provides name-based access to a bit in an msr based on the bits function as a flag.

Constructors

BitConf [Maybe l]

blank :: BitConfiguration l

an unnamed bit

typically used to set aside space for flow related operations, constants, or non-user accessible flags

blanks :: Int -> BitConfiguration l

Same as blank but for multiple blanks in order

named :: l -> BitConfiguration l

Defines a named bit that represents some flag in a Machine Status Register(MachineStatus).

names :: [l] -> BitConfiguration l

Same as named, but, provided a list of names, names the next n bits according to the provided list. (where n is the number of names provided)

getNumber :: Eq l => BitConfiguration l -> l -> Maybe Int

nameNumPairs :: FiniteBits r => BitConfiguration bsl -> r -> [(bsl, Bool)]