Boilerplate.Machines.State

Documentation

type MachineST l sl r = StateT (Machine l sl r)

Stateful Machine Definitions This should be an internal type only

Stateful Register Manipulation

setRegister :: (Monad m, Ord l) => l -> r -> MachineST l sl r m ()

getRegister :: (Monad m, Ord l) => l -> MachineST l sl r m (Maybe r)

modifyRegister :: (Monad m, Ord l) => l -> (r -> r) -> MachineST l sl r m ()

operate :: (Monad m, Ord l) => (r -> r -> r) -> l -> l -> l -> MachineST l sl r m r

Conducts an operation on two register values and places the result in a third register r_d

operateWithImmediate :: (Monad m, Ord l) => (r -> r -> r) -> l -> r -> l -> MachineST l sl r m ()

getMultipleRegs :: (Monad m, Ord l) => [l] -> MachineST l sl r m (Maybe [r])

Stateful Special Purpose Registers

setSpecialRegister :: (Monad m, Ord sl) => sl -> r -> MachineST l sl r m ()

getSpecialRegister :: (Monad m, Ord sl) => sl -> MachineST l sl r m (Maybe r)

modifySpecialRegister :: (Monad m, Ord sl) => sl -> (r -> r) -> MachineST l sl r m ()

setSPRBit :: (Monad m, FiniteBits r, Ord sl) => sl -> Int -> MachineST l sl r m ()

clearSPRBit :: (Monad m, FiniteBits r, Ord sl) => sl -> Int -> MachineST l sl r m ()

setWithConfigTo :: (Eq bsl, Monad m, FiniteBits r, Ord sl) => sl -> BitConfiguration bsl -> bsl -> Bool -> MachineST l sl r m ()

setWithConfig :: (Eq bsl, Monad m, FiniteBits r, Ord sl) => sl -> BitConfiguration bsl -> bsl -> MachineST l sl r m ()

clearWithConfig :: (Eq bsl, Monad m, FiniteBits r, Ord sl) => sl -> BitConfiguration bsl -> bsl -> MachineST l sl r m ()

testWithConfig :: (Eq bsl, Monad m, FiniteBits r, Ord sl) => sl -> BitConfiguration bsl -> bsl -> MachineST l sl r m (Maybe Bool)

Operations on Machine Status Registers

addWithCarry :: (FiniteBits r, Monad m, Ord l, Ord sl, Eq bsl, Num r, Ord r) => MSRBitAddress sl bsl -> r -> r -> l -> MachineST l sl r m Bool

applyCarry :: (Monad m, FiniteBits r, Ord sl, Eq bsl) => MSRBitAddress sl bsl -> Bool -> MachineST l sl r m ()

data MSRBitAddress sl bsl

Constructors

BitAddr sl (BitConfiguration bsl) bsl

Program Counter

branch :: (Monad m, Ord sl, Num r) => sl -> r -> MachineST l sl r m ()