Boilerplate.W32

Description

Documentation

zero :: W32

zero value (00000000 00000000 00000000 00000000)

one :: W32

one value (00000000 00000000 00000000 00000001)

Logic

and :: W32 -> W32 -> W32

Logical AND

or :: W32 -> W32 -> W32

Logical OR

not :: W32 -> W32

Logical NOT

(==) :: W32 -> W32 -> Bit

equality

xor :: W32 -> W32 -> W32

Conversions(Extensions)

signExtendW16 :: W16 -> W32

two's complement 16-bit sign extension

unsignedExtendW16 :: W16 -> W32

unsigned 16-bit extension

signExtendW8 :: W8 -> W32

two's complement byte sign extension

unsignedExtendW8 :: W8 -> W32

unsigned byte extension

backExtendW16 :: W16 -> W32

back-extends 16-bit data (useful for certain register operations)

example : 1010 0011 -> 1010 0011 0000 0000

leastSignificantByte :: W32 -> W8

the least significant byte

leastSignificantHalfWord :: W32 -> W16

the least significat hald-word

Arithmetic

add :: W32 -> W32 -> Bit -> (Bit, W32)

addition with carries

subtract :: W32 -> W32 -> Bit -> (Bit, W32)

subtraction with carries

signedCompare :: W32 -> W32 -> W32

subtraction-based comparison with sign-preservation

unsignedCompare :: W32 -> W32 -> W32

unsigned subtraction comparison

Value comparision

isNegative :: W32 -> Bit

test if value is negative under two's complement

isPositive :: W32 -> Bit

test if a value is positive or zero under two's complement

isZero :: W32 -> Bit

test if a value is zero

greaterThanZero :: W32 -> Bit

test if a value is greater than zero under two's complement

lessThanOrEqualToZero :: W32 -> Bit

test if a value is less than or equal to zero under two's complement

reverseSubtraction :: W32 -> W32 -> Bit -> (Bit, W32)

reverse subtraction

rb + not(ra) + carry

Bit Shifting

arithmeticShiftRight :: W32 -> Bit -> (Bit, W32)

logicalShiftRight :: W32 -> (Bit, W32)