Control.Monad.Resumption

Description

A resumption monad transformer, based on the formulation in the article Cheap (But Functional) Threads by William L. Harrison and Adam Procter.

Documentation

newtype ResT m a

Resumption monad transformer.

Constructors

ResT

Fields

deResT :: m (Either a (ResT m a))

runResT :: Monad m => ResT m a -> m a

Runs a resumptive computation to exhaustion, producing its final return value.

tick :: Monad m => ResT m ()

Waits until the next tick.