Control.Monad.Resumption.Reactive

Description

A reactive resumption monad transformer, based on the formulation in the article Cheap (But Functional) Threads by William L. Harrison and Adam Procter.

Documentation

newtype ReacT input output m a

Reactive resumption monad transformer.

Constructors

ReacT

Fields

deReacT :: m (Either a (output, input -> ReacT input output m a))

signal :: Monad m => output -> ReacT input output m input

Outputs its argument, then waits for the next input and returns it.

(<~>) :: Monad m => ReacT i o m a -> ReacT o i m b -> ResT m (Either a b)

Tennis operator.

runReacT :: Monad m => ReacT input output m a -> (output -> m input) -> m a

A basic runner function. Provide the ReacT and a handler in the underlying monad to run.