InstructionSet

Documentation

mbDefinition :: MachineDefinition String String Word32 IO ()

type MBInstruction m = Instruction String String Word32 m

data MBControl

Constructors

MBControl

Fields

delayFlag :: Bool

loadFlag :: Bool

storeFlag :: Bool

endOP :: ExecutionFunction String String Word32 IO MBControl

withDelay :: ExecutionFunction String String Word32 IO MBControl

instructionList :: InstructionSet String String Word32 IO ()

compareU :: Monad m => String -> String -> ExecutionFunction String String Word32 m ()

carryAddr :: MSRBitAddress String StatusBits

formA :: ArgumentsForm String

formB :: ArgumentsForm String

formBranchA :: ArgumentsForm String

formBranchB :: ArgumentsForm String

formUnconditionalBranch :: ArgumentsForm String

formUBranchWithLink :: ArgumentsForm String

parseMBReg :: Parser String

data StatusBits

Constructors

CC

VMS

VM

UMS

UM

PVR

EIP

EE

DCE

DZ

ICE

FSL

BIP

C

IE

BE

msrBitConf :: BitConfiguration StatusBits

testStatus :: Monad m => StatusBits -> MachineST l String Word32 m (Maybe Bool)

getRegisterValues :: (Monad m, Ord l) => [Argument l r] -> MachineST l sl r m (Maybe [r])

maybeCons :: Maybe x -> Maybe [x] -> Maybe [x]

typeA :: Monad m => (Word32 -> Word32 -> Word32) -> ExecutionFunction String String Word32 m ()

typeB :: Monad m => (Word32 -> Word32 -> Word32) -> ExecutionFunction String String Word32 m ()

addToPC :: Monad m => Word32 -> ExecutionFunction String String Word32 m ()

destFromAB :: Monad m => (Word32 -> Word32 -> Word32) -> ArgumentsForm String -> ExecutionFunction String String Word32 m ()

destFromABWithCarry :: Monad m => (Word32 -> Word32 -> Bool -> Word32) -> ArgumentsForm String -> ExecutionFunction String String Word32 m ()

getLiteralValuesForArgs :: (Monad m, Ord l) => [l] -> ExecutionFunction l sl r m (Map l r)

takes a list of argument names (ex: ra, rb, imm, etc.) and returns their values in terms of the current state

findVal :: (Monad m, Ord l) => Argument l r -> MachineST l sl r m (Maybe r)

getLiteralValueForArg :: (Monad m, Ord l) => l -> ExecutionFunction l sl r m (Maybe r)

operateOn

Arguments

:: (Monad m, Ord l)

=> (r -> r -> a)
The operation to perform on the two values

-> (l, l)
The values to be operated on (in-order of application)

-> ExecutionFunction l sl r m a

placeResultAt

Arguments

:: (Monad m, Ord l)

=> l
The register to set in terms of the description

-> r
The value to set the register to

-> ExecutionFunction l sl r m ()

testCarry :: Monad m => ExecutionFunction String String Word32 m Bool

setCarryTo :: Monad m => Bool -> ExecutionFunction String String Word32 m ()

overflow :: (Num r, FiniteBits r, Ord r) => r -> r -> r -> Bool

check if overflow will occur from addition

addRegular :: Monad m => (String, String) -> ExecutionFunction String String Word32 m ()

add :: Monad m => (String, String) -> Bool -> ExecutionFunction String String Word32 m Bool

keep :: Monad m => Bool -> ExecutionFunction String String Word32 m ()

carryOut :: Monad m => Bool -> ExecutionFunction String String Word32 m ()

addWithCarryS :: Monad m => (String, String) -> ExecutionFunction String String Word32 m Bool

addc :: (Num r, FiniteBits r, Ord r) => Bool -> r -> r -> (r, Bool)

withArg :: (Ord l, Monad m) => l -> (r -> ExecutionFunction l sl r m a) -> ExecutionFunction l sl r m a

with :: (Ord l, Monad m) => (r -> ExecutionFunction l sl r m a) -> l -> ExecutionFunction l sl r m a

with2 :: (Ord l, Monad m) => (r -> r -> ExecutionFunction l sl r m a) -> (l, l) -> ExecutionFunction l sl r m a

(.?.) :: (Ord l, Monad m) => l -> (r -> ExecutionFunction l sl r m Bool) -> ExecutionFunction l sl r m Bool

Check a predicate against an argument's state value

ifTrue :: (Monad m, Ord l) => ExecutionFunction l sl r m a -> Bool -> ExecutionFunction l sl r m ()

branchIf :: Monad m => String -> (Word32 -> Bool) -> String -> ExecutionFunction String String Word32 m ()