Interpreter

Description

Documentation

executeNext :: State MicroBlaze ()

nextPCAddress :: State MicroBlaze Address

isPositive :: Word32 -> Bool

isNegative :: Word32 -> Bool

lessThanOrEqualToZero :: Word32 -> Bool

exec :: Monad m => Ins -> StateT MicroBlaze m ()

Utility Functions and Data Types

addWithCarry :: (Num a, Ord a) => a -> a -> Bool -> (Bool, a)

type DelayFlag = Bool

Delay Flag for Branching

data BranchInput

Branch Input Type

Constructors

TypeA MBReg MBReg

TypeB MBReg Word16

data AbsoluteBranchInput

Absolute Branch Input Type

Constructors

AbsR MBReg

AbsI Word16

type Op = Word32 -> Word32 -> Word32

Isomorphism type

Generic Operator Execution

execTypeA

Arguments

:: Monad m

=> Op
Operator to apply to values

-> MBReg
Destination register

-> MBReg
Input Register A

-> MBReg
Input Register B

-> StateT MicroBlaze m ()

execute a type a instruction using basic operators

execTypeB

Arguments

:: Monad m

=> Op
Operator to apply to values

-> MBReg
Destination register

-> MBReg
Input Register A

-> Word16
Immediate Data

-> StateT MicroBlaze m ()

execute a type b instruction using basic operators

Branching

getBranchInputValue :: Monad m => BranchInput -> StateT MicroBlaze m Word32

SHOULD BE DEPRECATED

getBranchRegisterA :: Monad m => BranchInput -> StateT MicroBlaze m Word32

SHOULD BE DEPRECATED

absoluteBranch :: Monad m => AbsoluteBranchInput -> StateT MicroBlaze m ()

branch to an absolute address

branch

Arguments

:: Monad m

=> BranchInput
The branch input type (TypeA vs TypeB)

-> (MBWord -> Bool)
The predicate to decide branching

-> StateT MicroBlaze m ()

branch to a relative address

Addition

add :: Monad m => CarryFlag -> KeepFlag -> (MBReg, Either MBReg Word16) -> MBReg -> StateT MicroBlaze m ()

adding mechanism for MicroBlaze (likely can be deprecated)

sub :: Monad m => CarryFlag -> KeepFlag -> (MBReg, Either MBReg Word16) -> MBReg -> StateT MicroBlaze m ()

hardware subtraction (likely can be deprecated)

reverseSubtraction :: (Num a, Bits a, Ord a) => a -> a -> Bool -> (Bool, a)

type CarryFlag = Bool

Carry Flag for adder

type KeepFlag = Bool

Keep Flag for adder

Loading From Memory

data LoadSize

Constructors

LWord

LHalfWord

LByte

data ImmOrReg

Either-like datatype to differentiate TypeA and TypeB data

Constructors

Register MBReg

Immediate Word16

load

Arguments

:: Monad m

=> LoadSize
the size of the load operation (Byte, HalfWord, Word)

-> MBReg
Destination register for loaded data

-> MBReg
Register Offset 1

-> ImmOrReg
Register Offset 2

-> StateT MicroBlaze m ()

Loads data from memory, The two register offsets are added to obtain an address

data StoreSize

Constructors

SWord

SHalfWord

SByte

store :: Monad m => StoreSize -> MBReg -> MBReg -> ImmOrReg -> StateT MicroBlaze m ()

Special Purpose Registers

moveFromSRegister

Arguments

:: Monad m

=> MBReg
Destination Register

-> MBSReg
Special Purpose Register to Pull (MSR or RPC)

-> StateT MicroBlaze m ()

pulls either the MSR or PC register into given register

moveToSRegister

Arguments

:: Monad m

=> MBSReg
The register to alter

-> MBReg
The register containing the MSR Word

-> StateT MicroBlaze m ()

puts a Word into a special purpose register. Does not support updates to the program counter

link

Arguments

:: Monad m

=> MBReg
The destination register for the PC word

-> StateT MicroBlaze m ()

links the current program counter value into the specified register

delay :: Monad m => StateT MicroBlaze m ()

sets the delay flag in the machine status register

returnFrom :: Monad m => MBReg -> Word16 -> StateT MicroBlaze m ()

returns the pc from a break, interrupt, or subroutine

sext8 :: Monad m => MBReg -> MBReg -> StateT MicroBlaze m ()

sext16 :: Monad m => MBReg -> MBReg -> StateT MicroBlaze m ()

shiftRightArithmetic :: Monad m => CarryFlag -> MBReg -> MBReg -> StateT MicroBlaze m ()

shiftRightLogical :: Monad m => MBReg -> MBReg -> StateT MicroBlaze m ()

logicalShiftRight :: Bits a => a -> (Bool, a)

arithmeticShiftRight :: FiniteBits a => a -> Bool -> (Bool, a)

leastSignificantHalfWord :: Word32 -> Word16

leastSignificantByte :: Word32 -> Word8

maybeSignExtend :: (Integral a, Integral b, FiniteBits a, FiniteBits b) => a -> Maybe b

showFiniteBits :: FiniteBits a => a -> String

backExtend :: (Integral a, Integral b, FiniteBits a, FiniteBits b) => a -> Maybe b