MachineState

Description

Microblaze Machine type

data MicroBlaze

Full MicroBlaze Register Profile

Constructors

MicroBlaze MBRegisters RPC RMSR InstructionBuffer

type MBWord = W32

Standard word size for MicroBlaze

newMicroBlaze :: MicroBlaze

Creates a zero-initialized MicroBlaze controller

Special-Purpose Registers

type RPC = MBWord

Program Counter

getRPC :: State MicroBlaze MBWord

Gets the RPC from a MicroBlaze State

setRPC :: MBWord -> State MicroBlaze ()

Sets the RPC in a MicroBlaze State

Machine Status Register

setMSRBit

Arguments

:: MachineStatusBit
Machine Status Flag to set

-> Bit
The value to set it to

-> State MicroBlaze ()

Sets a specified bit on the machine status register

getMSRBit :: MachineStatusBit -> State MicroBlaze Bit

Gets the value of a specified MSR Bit

pullMSR :: State MicroBlaze W32

Pulls the enture Machine Status Register as a 32-bit Word

pushMSR :: W32 -> State MicroBlaze ()

Pushes an enture MSR as a 32-bit word

General-Purpose Registers

data MBRegisters

The register store for the Microblaze

emptyRegisters :: MBRegisters

zero-initialized register bank

Register Manipulation

getRegister :: MBReg -> State MicroBlaze MBWord

gets the value at a specified register

setRegister :: MBReg -> MBWord -> State MicroBlaze ()

sets the value at a specified register

readRegister :: MBReg -> MBRegisters -> MBWord

read a given register

Memory Managment

type Address = W32

Loading

loadWord :: W32 -> State MicroBlaze W32

loads a 32-bit word from memory

NOT YET IMPLEMENTED

loadHalfWord :: W32 -> State MicroBlaze W16

loads a 16-bit half-word from memory

NOT YET IMPLEMENTED

loadByte :: W32 -> State MicroBlaze W8

loads a byte from memory

NOT YET IMPLEMENTED

Storing

storeWord :: W32 -> W32 -> W32 -> State MicroBlaze ()

stores a 32-bit word in memory

NOT YET IMPLEMENTED

storeHalfWord :: W16 -> W32 -> W32 -> State MicroBlaze ()

stores a 16-bit half-word in memory

NOT YET IMPLEMENTED

storeByte :: W8 -> W32 -> W32 -> State MicroBlaze ()

stores a 8-bit byte in memory

NOT YET IMPLEMENTED

Instruction Buffer

fetch :: Address -> State MicroBlaze W32

pushInstructionBuffer :: InstructionBuffer -> State MicroBlaze ()

pullDecode :: State MicroBlaze (Maybe W32)

pullExec :: State MicroBlaze (Maybe Ins)