MachineState.MachineStatusRegister

Description

Machine Status Register

data RMSR

emptyRMSR :: RMSR

a zero-initialized (All False) MSR

getMSRWord :: RMSR -> Word32

gets the status register in the form of a 32-bit word

setMSRWord :: Word32 -> RMSR

sets the status register from a 32-bit MSR word

Status Flags

data MachineStatusBit

Various status flags in the Machine State Register

Constructors

CarryCopy

DataCacheEnable

DivisionByZero

InstructionCacheEnable

FSLError

BreakInProgress

Carry

InterruptEnable

BuslockEnable

DelayEnable

getStatus :: MachineStatusBit -> RMSR -> Bool

Get the status of a specified flag

setStatus :: MachineStatusBit -> Bool -> RMSR -> RMSR

sets the machine status bit indicated to the desired boolean value