Boilerplate

Description

Bit Data

data Bit

Standard Bit Datatype Supporting functions can be found in Boilerplate.Bit

Constructors

C

S

Byte and Standard Size Data

data W8

Standard Byte defintion Supporting functions can be found in Boilerplate.W8

Constructors

W8 Bit Bit Bit Bit Bit Bit Bit Bit

data W16

16-bit data

Constructors

W16 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

data W32

32-bit data Supporting functions can be found in Boilerplate.W32

Constructors

W32 W8 W8 W8 W8

Unusual/Rare-Sized Bit Data

data W5

5-bit data

Constructors

W5 Bit Bit Bit Bit Bit

data W6

6-bit data

Constructors

W6 Bit Bit Bit Bit Bit Bit

data W11

11-bit data

Constructors

W11 Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

maybeSignExtend :: (Integral a, Integral b, FiniteBits a, FiniteBits b) => a -> Maybe b

showFiniteBits :: FiniteBits a => a -> String