Boilerplate.Machines

Machines

data Machine l sl r

A machine of inert(stateless) registers

machine :: FiniteBits r => Int -> [String] -> Machine String String r

Register Banks

data RegisterBank l r

A register bank with labels l and registers r

newRegisterBank

Arguments

:: (FiniteBits r, Ord l)

=> Int
Size of the bank

-> (Int -> l)
Function to generate labels

-> RegisterBank l r

Creates a new register bank using a generator functions to produce label l

standardRegisterBank :: FiniteBits r => Int -> RegisterBank String r

produces a register bank with numeric labels proceeded by the character r. For example the first register is "r0" Important to note that this produces a strange order from labels being Strings

Special Purpose Register Banks

type SpecialRegisters sl r = RegisterBank sl r

specialFromList :: (FiniteBits r, Ord l) => [l] -> SpecialRegisters l r

Examples

MicroBlaze

mb :: Machine String String Word32