Interpreter

Exec Function

exec :: Ins -> State MicroBlaze ()

note that all exec does is process the given instructions as it can given the machine state information. Delays and other temporal actions must be handled at a different level of the program

Functions * Utility Functions and Data Types

type CarryFlag = Bool

type KeepFlag = Bool

type DelayFlag = Bool

data BranchInput

Constructors

TypeA MBReg MBReg

TypeB MBReg W16

data AbsoluteBranchInput

Constructors

AbsR MBReg

AbsI W16

type Op = W32 -> W32 -> W32

execTypeA :: Op -> MBReg -> MBReg -> MBReg -> State MicroBlaze ()

execTypeB :: Op -> MBReg -> MBReg -> W16 -> State MicroBlaze ()

link :: MBReg -> State MicroBlaze ()

delay :: State MicroBlaze ()

getBranchInputValue :: BranchInput -> State MicroBlaze W32

getBranchRegisterA :: BranchInput -> State MicroBlaze W32

absoluteBranch :: AbsoluteBranchInput -> State MicroBlaze ()

branch :: BranchInput -> (MBWord -> Bit) -> State MicroBlaze ()

add :: CarryFlag -> KeepFlag -> (MBReg, Either MBReg W16) -> MBReg -> State MicroBlaze ()

data LoadSize

Constructors

LWord

LHalfWord (W16 -> W32)

LByte (W8 -> W32)

data ImmOrReg

Constructors

Register MBReg

Immediate W16

load :: LoadSize -> MBReg -> MBReg -> ImmOrReg -> State MicroBlaze ()

moveFromSRegister :: MBReg -> MBSReg -> State MicroBlaze ()

moveToSRegister :: MBSReg -> MBReg -> State MicroBlaze ()