Interpreter

Documentation

type Op = W32 -> W32 -> W32

execTypeA :: Op -> MBReg -> MBReg -> MBReg -> State MicroBlaze ()

execTypeB :: Op -> MBReg -> MBReg -> W16 -> State MicroBlaze ()

exec :: Ins -> State MicroBlaze ()

note that all exec does is process the given instructions as it can given the machine state information. Delays and other temporal actions must be handled at a different level of the program

type CarryFlag = Bool

type KeepFlag = Bool

type DelayFlag = Bool

branch :: DelayFlag -> (MBReg, Either MBReg W16) -> (MBWord -> Bit) -> State MicroBlaze ()

add :: CarryFlag -> KeepFlag -> (MBReg, Either MBReg W16) -> MBReg -> State MicroBlaze ()