MachineState.MachineStatusRegister

Documentation

data RMSR

the machine status register

Constructors

RMSR Bit Bit Bit Bit Bit Bit Bit Bit Bit Bit

emptyRMSR :: RMSR

data MachineStatusBit

Constructors

CarryCopy

DataCacheEnable

DivisionByZero

InstructionCacheEnable

FSLError

BreakInProgress

Carry

InterruptEnable

BuslockEnable

DelayEnable

getStatus :: MachineStatusBit -> RMSR -> Bit

setStatus :: MachineStatusBit -> Bit -> RMSR -> RMSR

sets the machine status bit indicated to the desired boolean value

getMSRWord :: RMSR -> W32

gets the status register in the form of a 32-bit word

setMSRWord :: W32 -> RMSR